update UI and refactor functions
