+++ src/components/InputDisplay.tsx
@@ -1,56 +1,114 @@
 import styled from 'styled-components';
-import { IonInput, IonSelect, IonSelectOption, IonLabel } from '@ionic/react';
-// import { useKeyboardState } from '@ionic/react-hooks/keyboard';
+import {
+  IonInput,
+  IonSelect,
+  IonSelectOption,
+  IonButton,
+  IonIcon,
+  IonText,
+} from '@ionic/react';
+import { closeCircleOutline } from 'ionicons/icons';
 import percentages from '../utils/data/percentages';
 
-interface Props {
-  isOpen: any;
-  keyboardHeight: any;
-}
+// Redux
+import type { RootState } from '../redux/app/store';
+import { useSelector, useDispatch } from 'react-redux';
+import {
+  setInputWeight,
+  setPlateData,
+  setPercentage,
+  setFinalWeight,
+  reset,
+} from '../redux/features/plateSlice';
+import { getDataFromInput } from '../utils/functions';
 
-const InputDisplay: React.FC = () => {
-  // const { isOpen, keyboardHeight } = useKeyboardState();
+const InputDisplay: React.FC<any> = () => {
+  const dispatch = useDispatch();
+  const { inputWeight, percentage, finalWeight } = useSelector(
+    (state: RootState) => state.plate
+  );
 
-  return null;
+  const onInputWeightChange = (e: any) => {
+    dispatch(setInputWeight(e.target.value!));
+    if (e.target.value === '') {
+      dispatch(reset());
+    }
+  };
 
-  // return (
-  //   <Wrapper>
-  //     <FormControl isOpen={isOpen} keyboardHeight={keyboardHeight}>
-  //       <IonLabel>Weight (lbs)</IonLabel>
-  //       <IonInput type='number' inputMode='numeric' placeholder='Weight' />
-  //     </FormControl>
-  //     <FormControl isOpen={isOpen} keyboardHeight={keyboardHeight}>
-  //       <IonLabel>Percentage</IonLabel>
-  //       <IonSelect value={100} placeholder='Percentage'>
-  //         {percentages.map((percentage) => (
-  //           <IonSelectOption key={percentage.value} value={percentage.value}>
-  //             {percentage.label}
-  //           </IonSelectOption>
-  //         ))}
-  //       </IonSelect>
-  //     </FormControl>
-  //   </Wrapper>
-  // );
+  const handleOnClick = () => {
+    const data = getDataFromInput(inputWeight, percentage);
+    dispatch(setFinalWeight(data.finalWeight));
+    dispatch(setPlateData(data.plateData));
+  };
+
+  const handlePercentageChange = (value: string) => {
+    dispatch(setPercentage(value));
+
+    const data = getDataFromInput(inputWeight, value);
+    dispatch(setFinalWeight(data.finalWeight));
+    dispatch(setPlateData(data.plateData));
+  };
+
+  return (
+    <Wrapper>
+      <InputWrapper>
+        <Input
+          value={inputWeight}
+          type='text'
+          inputMode='numeric'
+          placeholder='LBS'
+          size={3}
+          onIonChange={(e: any) => onInputWeightChange(e)}
+        ></Input>
+        <IonButton onClick={() => dispatch(reset())} fill='clear'>
+          <IonIcon icon={closeCircleOutline} />
+        </IonButton>
+      </InputWrapper>
+      <StyledIonSelect
+        interface='popover'
+        value={percentage}
+        placeholder='Percentage'
+        onIonChange={(e) => handlePercentageChange(e.target.value)}
+      >
+        {percentages.map((percentage) => (
+          <IonSelectOption key={percentage.value} value={percentage.value}>
+            {percentage.label}
+          </IonSelectOption>
+        ))}
+      </StyledIonSelect>
+      <IonButton onClick={() => handleOnClick()}>Calc</IonButton>
+      {Number(inputWeight) !== finalWeight && (
+        <IonText color='warning'>
+          <p>
+            Final Weight: <span>{finalWeight}</span>{' '}
+          </p>
+        </IonText>
+      )}
+    </Wrapper>
+  );
 };
 
 export default InputDisplay;
 
-const Wrapper = styled.div`
+const InputWrapper = styled.div`
   display: flex;
-  flex-direction: row;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
-  max-width: 500px;
-  margin: auto auto;
-  margin-bottom: 10px;
-  text-align: center;
-  /* gap: 10px; */
+  margin-left: 65px;
 `;
 
-const FormControl = styled.div<Props>`
-  width: 100%;
-  margin-top: ${(props) =>
-    props.isOpen ? `${props.keyboardHeight * -1 - 200}px` : null};
-  background-color: --ion-background-color;
-  z-index: 100;
+const Wrapper = styled.div`
+  margin: 0 auto;
+  text-align: center;
+`;
+
+const StyledIonSelect = styled(IonSelect)`
+  && {
+    width: 150px;
+    margin: 0 auto;
+  }
+`;
+
+const Input = styled(IonInput)<{ value: string }>`
+  font-size: 110px;
 `;
