--- src/components/InputDisplay2.css
@@ -1,3 +1,0 @@
-ion-input {
-  --color: #333;
-}
